Moved form classes to new file forms.py
