+++ app/forms.py
@@ -1,0 +1,15 @@
+from flask_wtf import FlaskForm
+from wtforms import SubmitField, StringField, PasswordField
+from wtforms.validators import DataRequired
+
+class RegistrationForm(FlaskForm):
+    email = StringField('Email', validators=[DataRequired()])
+    username = StringField('Username', validators=[DataRequired()])
+    password = PasswordField('Password', validators=[DataRequired()])
+    confirm_password = PasswordField('Confirm Password', validators=[DataRequired()])
+    submit = SubmitField('Register')
+
+class LoginForm(FlaskForm):
+    username = StringField('Username', validators=[DataRequired()])
+    password = PasswordField('Password', validators=[DataRequired()])
+    submit = SubmitField('Login')
